# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESG Hub API",
    "version": "1.0.0",
    "description": "Public API for the Ascent Partners Foundation ESG Knowledge Hub. Access 354 ESG articles and 244 curated external resources covering Environmental, Social, and Governance topics.",
    "contact": {
      "name": "ESG Hub Developers",
      "url": "https://esg-hub.ascent.partners/developers"
    },
    "license": {
      "name": "CC BY-SA 4.0",
      "url": "https://creativecommons.org/licenses/by-sa/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://esg-hub.ascent.partners",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/pages/{id}/facets": {
      "patch": {
        "summary": "Update page facets",
        "description": "Update taxonomy facets on an existing page. Requires Bearer token authorization.",
        "operationId": "updateFacets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Page record ID (e.g. page:abc123)"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "facets": {
                    "type": "object",
                    "properties": {
                      "topic": { "type": "array", "items": { "type": "string" } },
                      "industry": { "type": "array", "items": { "type": "string" } },
                      "framework": { "type": "array", "items": { "type": "string" } },
                      "jurisdiction": { "type": "array", "items": { "type": "string" } },
                      "stakeholder": { "type": "array", "items": { "type": "string" } },
                      "content_type": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated page with facets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Page not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "ESG Hub Write Token (required for POST/PATCH endpoints)"
      }
    }
  },
...
}